backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
